requirements.txt
langchain
langchain-openai
langchain-astradb
datasets
pypdf
python-dotenv
flask
langchain-community
uvicorn
pandas
-e .
